api call tested & working first mock for UX working property rating in Model commented, needs refactor
