+++ SwiftUIList/SwiftUIList/MainView/ViewModel/ShowListViewModel.swift
@@ -12,7 +12,6 @@ class ShowListViewModel: ObservableObject {
     @Published var showList = [ShowListModel]()
     
     func callToApi() async {
-        
         guard let url = URL(string: Constants.showListURL) else { return }
         
         do {
@@ -20,6 +19,7 @@ class ShowListViewModel: ObservableObject {
             if let response = try? JSONDecoder().decode([ShowListModel].self, from: data) {
                 DispatchQueue.main.async {
                     self.showList = response
+                    print(response)
                 }
             }
         } catch {
